requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
drf-writable-nested==0.4.2
mysqlclient==1.3.12
Pillow==5.1.0
pytz==2018.4